+++ owlapy/owl_axiom.py
@@ -3,7 +3,7 @@ from abc import ABCMeta, abstractmethod
 
 from typing import TypeVar, List, Optional, Iterable, Generic, Final
 from .owl_property import OWLDataPropertyExpression, OWLObjectPropertyExpression
-from .owlobject import OWLObject, OWLEntity
+from .owl_object import OWLObject, OWLEntity
 from .types import OWLDatatype, OWLDataRange
 from .meta_classes import HasOperands
 from .owl_property import OWLPropertyExpression, OWLProperty
